+++ triplea/the_private_backyard2.py
@@ -20,8 +20,12 @@ import traceback
 import os
 
 from triplea.service.repository.pipeline_core import move_state_forward
+from triplea.service.repository.pipeline_flag import go_affiliation_mining, go_extract_topic
 
 if __name__ == "__main__":
     pass
-    # export_triplea_csvs_in_relational_mode_save_file('export.csv')
-    move_state_forward(2)
+    export_triplea_csvs_in_relational_mode_save_file('export.csv',
+                                                     proccess_bar=True)
+    # move_state_forward(2)
+    # go_affiliation_mining()
+    # go_extract_topic(proccess_bar=True)